refactor table view
