+++ demo-app/src/pages/platform-devices/TableView/index.js
@@ -1,0 +1,1 @@
+export * from './TableView';
